Два коэффициента стоимости, вкладки деревьев и окно карточки
1. Числа выводятся с неразрывным пробелом через каждые три знака —
   и на доске, и в панелях столбцов, и в списке технологий. Пробел
   именно неразрывный, иначе стоимость рвётся по переносу строки.
   Поле среднего над столбцом стало текстовым, значение разбирается
   обратно с пробелами.

2. Коэффициент стоимости разделён на два: cost_step_lane (переход
   к следующему столбцу внутри эпохи, по умолчанию 1.1) и cost_step_era
   (переход через границу эпохи, по умолчанию 1.5). Множитель столбца
   считается проходом слева направо, поэтому внутри эпохи цена растёт
   полого, а на смене эпохи делает заметный шаг. Оба коэффициента выведены
   в закреплённую панель над деревьями и меняются вручную — правка сразу
   пересчитывает все эпохи. Разброс внутри столбца берётся от меньшего
   из коэффициентов, чтобы диапазоны соседних столбцов не пересекались.

3. Деревья выводятся не подряд, а по вкладкам. Скрытую доску нельзя
   обмерить, поэтому показанное дерево перерисовывается при переключении;
   переход к карточке с чужой вкладки сам открывает нужное дерево.

4. Доска прокручивается в обе стороны внутри своей рамки и таскается
   мышью за фон. По карточкам и кнопкам не тянем, чтобы не мешать кликам
   и правке связей.

5. На карточке появилась кнопка «i»: открывает окно с картинкой, описанием,
   исторической справкой и списками ближайших основ и последователей —
   каждый со стоимостью в скобках и кликабельный. Клик по соседу
   перезагружает окно на него, есть шаг назад и переход на карточку
   в дереве. Описания и справки на доску не отдаются (623 записи
   по нескольку абзацев) — окно забирает их отдельным запросом.

Стоимости базовой версии пересчитаны по новым коэффициентам: самая дорогая
технология теперь стоит 2 310 442 вместо полутора миллиардов.

+++ app/views/technologies.php
@@ -72,7 +72,7 @@ $f = $filter;
       <td><span class="swatch" style="background: <?= h($row['branch_color']) ?>"></span><?= h($row['branch_name']) ?></td>
       <td><?= h($row['era_name']) ?></td>
       <td><?= $row['is_standard'] ? 'стандартный' : '<span class="tag">вне набора</span>' ?></td>
-      <td><?= $row['base_cost'] !== null ? (int) $row['base_cost'] : '<span class="hint">по столбцу</span>' ?></td>
+      <td><?= $row['base_cost'] !== null ? money($row['base_cost']) : '<span class="hint">по столбцу</span>' ?></td>
       <td><?= (int) $row['effect_count'] ?></td>
     </tr>
   <?php endforeach; ?>
